requirements.txt
flask>=3.0.0
flask-login>=0.6.3
flask-socketio>=5.3.6
python-dotenv>=1.0.0
dalybms
pyserial>=3.5
APScheduler>=3.10.4
